requirements.txt
anthropic>=0.39.0
openai>=1.0.0
pdfplumber>=0.11.0
genanki>=0.13.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
